requirements.txt
annotated-types==0.7.0
anyio==4.9.0
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.3
fastapi==0.115.12
h11==0.16.0
idna==3.10
joblib==1.5.1
nltk==3.9.1
numpy==2.2.6
pdfminer.six==20250327
pdfplumber==0.11.6
pillow==11.2.1
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pypdfium2==4.30.1
python-multipart==0.0.20
regex==2024.11.6
scikit-learn==1.6.1
scipy==1.15.3
sniffio==1.3.1
starlette==0.46.2
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.2
uvicorn==0.34.2
